requirements.txt
coloredlogs==15.0.1
dataclasses-json==0.6.7
humanfriendly==10.0
marshmallow==3.23.1
mypy-extensions==1.0.0
packaging==24.2
psutil==6.1.0
pyreadline3==3.5.4
python-dotenv==1.0.1
typing-inspect==0.9.0
typing_extensions==4.12.2
